requirements.txt
auth==0.5.3
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
esdk-obs-python==3.22.2
eventlet==0.33.3
falcon==3.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
MarkupSafe==2.1.3
mongoengine==0.27.0
pycryptodome==3.10.1
pymongo==4.4.0
python-dotenv==1.0.0
requests==2.31.0
signer==0.0.1
six==1.16.0
urllib3==2.0.3
Werkzeug==2.3.6
xmltodict==0.13.0
